# backend/docs/swagger.yaml
basePath: /
definitions:
...
  http.PaginationInfo:
    properties:
      is_pager:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  http.ProjectListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/http.ProjectResp'
        type: array
      pagination:
        $ref: '#/definitions/http.PaginationInfo'
    type: object
  http.ProjectResp:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
    type: object
  http.R:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  http.RProject:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.ProjectResp'
      msg:
        type: string
    type: object
  http.RProjectList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.ProjectListResp'
      msg:
        type: string
    type: object
...
  http.RTenant:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.TenantResp'
      msg:
        type: string
    type: object
...
  http.TenantResp:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      plan:
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: 企业级 SQL to API 智能网关平台管理接口
  title: OwlApi Control Plane
  version: "1.0"
paths:
  /api/v1/tenants/{slug}/query/{path}:
    post:
      consumes:
      - application/json
      operationId: executeQuery
      parameters:
      - description: 租户 slug
        in: path
        name: slug
        required: true
        type: string
      - description: API 路径
        in: path
        name: path
        required: true
        type: string
      - description: 查询参数（由端点 param_defs 定义）
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: 执行动态 API 查询
      tags:
      - query
...
  /v1/tenants/{slug}:
    delete:
      operationId: deleteTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 删除租户 (SuperAdmin)
      tags:
      - tenant
    get:
      operationId: getTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenant'
      security:
      - BearerAuth: []
      summary: 获取租户详情
      tags:
      - tenant
    put:
      consumes:
      - application/json
      operationId: updateTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          properties:
            name:
              type: string
            plan:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenant'
      security:
      - BearerAuth: []
      summary: 更新租户 (SuperAdmin)
      tags:
      - tenant
...
  /v1/tenants/{slug}/projects:
    get:
      operationId: listProjects
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10）
        in: query
        name: size
        type: integer
      - description: 是否分页，0=返回全部（默认1）
        in: query
        name: is_pager
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RProjectList'
      security:
      - BearerAuth: []
      summary: 获取项目列表
      tags:
      - project
    post:
      consumes:
      - application/json
      operationId: createProject
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目信息
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RProject'
      security:
      - BearerAuth: []
      summary: 创建项目
      tags:
      - project
  /v1/tenants/{slug}/projects/{projectId}:
    delete:
      operationId: deleteProject
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 删除项目
      tags:
      - project
    get:
      operationId: getProject
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RProject'
      security:
      - BearerAuth: []
      summary: 获取项目详情
      tags:
      - project
    put:
      consumes:
      - application/json
      operationId: updateProject
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RProject'
      security:
      - BearerAuth: []
      summary: 更新项目
      tags:
      - project
...
  /v1/tenants/{slug}/projects/{projectId}/openapi.json:
    get:
      operationId: exportOpenApi
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 导出项目 OpenAPI 规范
      tags:
      - project
  /v1/tenants/{slug}/query/test:
    post:
      consumes:
      - application/json
      operationId: testQuery
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 测试参数
        in: body
        name: body
        required: true
        schema:
          properties:
            endpoint_id:
              type: integer
            ignore_scripts:
              type: boolean
            params:
              type: object
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 测试执行 API 端点
      tags:
      - query
...
securityDefinitions:
  BearerAuth:
    description: '格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"